requirements.txt
discord.py==2.3.2
pandas>=2.0.0
jaconv>=0.3.0
aiohttp>=3.8.0
pillow>=9.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0